ss: fix host build
We had updated it for cross compilation based on arm rsk, but their patch was wrong,
ssl's configure is not automake and does not take --build and --host.

Maybe cross compilation it will just work after this patch due to gcc being
identified by CC, OS appears to be identified by uname, so it will work "by
coincidence"" for Linux build hosts.